% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deg_analysis.R
\name{deg_analysis}
\alias{deg_analysis}
\title{deg_analysis}
\usage{
deg_analysis(count_data, design, contrastmat)
}
\arguments{
\item{count_data}{A dataframe of RNA-seq count data (filtered)}

\item{design}{A dataframe containing the study explanatory variables.}

\item{contrastmat}{A contrast matrix containing the sample groups which are
being compared}
}
\value{
Table of significant DEG's between sample groups.
}
\description{
Performs differentially expressed genes (DEG) analysis using functions in
edgeR.

Generalized linear modeling is applied to response variables (genes here),
which are not normally distributed.

The function needs to be supplied with the design and contrast matrix that
corresponds to the study.
}
\examples{
library(EnrichTool)
deg_analysis(count_data, design, contrastmat)

}
