% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export_to_excel.R
\name{export_to_excel}
\alias{export_to_excel}
\title{export_to_excel}
\usage{
export_to_excel(deg_res, file_name = "deg_results.xlsx")
}
\arguments{
\item{deg_res}{Result from DEG-analysis}

\item{file_name}{Name of file.}
}
\value{
Returns an excel file with DEG's.
}
\description{
Writes the results from differentially expressed genes (DEG) analysis to Excel.
}
\examples{
library(EnrichTool)
export_to_excel(deg_res, file_name = "my_deg_results.xlsx")

}
