% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_RNA.R
\name{import_RNA}
\alias{import_RNA}
\title{import_RNA}
\usage{
import_RNA(file, cutoff = 1)
}
\arguments{
\item{file}{RNA-seq count file.}

\item{cutoff}{Cutoff value for filtering rows with low expression}
}
\value{
Filtered RNA-seq count data.
}
\description{
Imports RNA-seq count data and performs filtering on log2 CPM
}
\examples{
library(EnrhichTool)
import_RNA("path/to/file.txt", cutoff = 1)

}
