% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/importRna.R
\name{importRna}
\alias{importRna}
\title{importRna}
\usage{
importRna(file, cutoff)
}
\arguments{
\item{file}{RNA-seq count file.}

\item{cutoff}{Cutoff value for filtering rows with low expression}
}
\value{
Filtered RNA-seq count data.
}
\description{
Imports RNA-seq count data and performs filtering on log2 CPM
}
