% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ora_plot.R
\name{ora_plot}
\alias{ora_plot}
\title{ora_plot}
\usage{
ora_plot(ora_res, plots = c("dotplot", "cnetplot"))
}
\arguments{
\item{ora_res}{A list with result from ORA}

\item{plots}{The string of desired plot.}
}
\value{
Plots that visualizes the results from ORA
}
\description{
Plots the result from over representation analysis from KEGG and GO.

Choose to visualize ora results with dotplot and/or cnetplot.
}
\examples{
library(EnrichTool)
ora_plot(ora_res, plots = "dotplot")

}
