% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ora_res.R
\name{ora_res}
\alias{ora_res}
\title{ora_res}
\usage{
ora_res(gene_data, key_type = "ENSEMBL")
}
\arguments{
\item{gene_data}{Dataframe where rownames are genes.}

\item{key_type}{Key type of the genes in the dataframe.}
}
\value{
Ora results.
}
\description{
Performs over-representation analysis on set of genes.

The "gene_data" is a dataframe where rownames are genes. Typically the
output of a differentially expressed genes analysis. The rownames are used
to convert to entrez id before enrichment.
}
\examples{
library(EnrichTool)
ora_res(deg_res, key_type = "ENSEMBL")

}
